requirements.txt
streamlit
pymongo
bcrypt
python-jose
pdfplumber
python-docx
fastapi
uvicorn
python-dotenv
pydantic
requests
python-multipart
email-validator
nltk
spacy
scikit-learn
